backend/pyproject.toml
[project]
name = "skillnexus-backend"
version = "0.1.0"
description = "SkillNexus - AI-Powered Enterprise L&D Platform Backend"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Pydantic & Settings
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Auth & Security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",

    # Redis (stateful refresh tokens)
    "redis>=5.2.0",

    # LangChain & LLM Providers
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.0",

    # PDF Parsing
    "pdfminer-six>=20231228",

    # Utilities
    "python-dateutil>=2.9.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
]
